Move AgentTrainer into preference_comparisons.py
It's become pretty specific to preference comparisons
now and if you just want to train on a custom reward_fn
without collecting trajectories, that's very easy to do inline.
